+++ src/main/scala/play/api/cache/redis/impl/RedisCache.scala
@@ -49,7 +49,7 @@ private[impl] class RedisCache[Result[_]](redis: RedisConnector, builder: Builde
       redis.append(key, value).flatMap { result =>
         // if the new string length is equal to the appended string, it means they should equal
         // when the finite duration is required, set it
-        if (result === value.length && expiration.isFinite) redis.expire(key, expiration) else Future.successful[Unit](())
+        if (result === value.length.toLong && expiration.isFinite) redis.expire(key, expiration) else Future.successful[Unit](())
       }.recoverWithDone
     }
